+++ src/Data/TimeSeries.hs
@@ -1,9 +1,11 @@
 -- | TimeSeries library
 module Data.TimeSeries
     ( module Data.TimeSeries.Series
-    , module Data.TimeSeries.CSVReader
+    , module Data.TimeSeries.IO.CSVReader
+    , module Data.TimeSeries.IO.CSVWriter
     ) where
 
 
 import Data.TimeSeries.Series
-import Data.TimeSeries.CSVReader
+import Data.TimeSeries.IO.CSVReader
+import Data.TimeSeries.IO.CSVWriter